# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presente profe API 2024",
        "contact": {
            "name": "Benjamin Mora",
            "email": "bej.mora@profesor.duoc.cl"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://www.presenteprofe.test/",
            "description": "Production"
        },
        {
            "url": "https://presenteprofe.test/",
            "description": "Ambiente de desarrollo"
        }
    ],
    "paths": {
        "/api/v1/auth": {
            "post": {
                "tags": [
                    "Auth:Autenticación"
                ],
                "summary": "Iniciar sesión y generar token de autenticación",
                "operationId": "956b567bea82b5707296567cb2bf27c6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "password"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "mypassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "correo": {
                                                    "type": "string",
                                                    "example": "example@example.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "auth": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error en las credenciales",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalles del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth:Usuarios"
                ],
                "summary": "Obtener mi informacion de usuario",
                "operationId": "8bb633895a24921f8bb59f4db5a8e17e",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "description": "Correo electrónico del usuario",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "correo@gmail.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informacion de usuario obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "run": {
                                                    "type": "string",
                                                    "example": "12345678-9"
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Juan"
                                                },
                                                "apellido": {
                                                    "type": "string",
                                                    "example": "Pérez"
                                                },
                                                "nombre_completo": {
                                                    "type": "string",
                                                    "example": "Juan Pérez"
                                                },
                                                "correo": {
                                                    "type": "string",
                                                    "example": "juan.perez@example.com"
                                                },
                                                "perfil": {
                                                    "type": "string",
                                                    "example": "estudiante"
                                                },
                                                "img": {
                                                    "type": "string",
                                                    "example": "http://example.com/path/to/image.jpg"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/usuarios": {
            "post": {
                "tags": [
                    "Auth:Autenticación"
                ],
                "summary": "Registrar un nuevo usuario mediante código de invitación",
                "operationId": "6b4700583a55930d0b80dd98f98f2a71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "codigo": {
                                        "type": "string",
                                        "example": "presenteprofe"
                                    },
                                    "run": {
                                        "type": "string",
                                        "example": "12345678-9"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "correo@gmail.com"
                                    },
                                    "perfil": {
                                        "type": "string",
                                        "enum": [
                                            "docente",
                                            "estudiante"
                                        ],
                                        "example": "docente|estudiante"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación, falta el perfil"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Código de invitación no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Código de invitación no válido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/recuperar": {
            "post": {
                "tags": [
                    "Auth:Autenticación"
                ],
                "summary": "Recuperar la contraseña del usuario",
                "operationId": "1d8ea1cade6e2980537e5793ecbd2e53",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "correo@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Solicitud de recuperación de contraseña enviada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud de recuperación de contraseña enviada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocurrió un error durante la recuperación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un error durante la recuperación"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Descripción del error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}